resume_screening/requirements.txt
PyPDF2
python-docx
spacy
scikit-learn
openai
flask